Add support for github enterprise
GitHub enterprise just requires setting a git_host to something other
than the default of 'github.com'. If this host is set to something else,
perform an enterprise_login() instead of a login(). All other activity
appears to be the same.

A more complete change would be allowing clone/push using the API key
instead of the ssh key, which would reduce the amount of credentials one
must manage on the zuul host.

This change does not add tests.