Update repo origin on update
There are cases where the on-disk configuration for a git remote can
drift from the remote URL we get from zuul config. Without updating it,
fetch attempts will fail as only the in-repo configuration is
considered. This is particularly bad when we're using short lived access
tokens to access repositories.

Change-Id: Id4042d95a8738e8f91164c7ba38a5704dcd2cb86