Adds github triggering from status updates
This adds support for triggering on github status updates.

The status matching logic was moved out of the ChangeishFilter
and into BaseFilter so it can also be used by EventFilters.  Config
schema for the github trigger has been updated to accept a list of
statuses, in the "github_user:context:status" format.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>