Move PR review states to constants, fix broken test
PR #18 adjusted the review state strings we are keying on but did
not update the corresponding test case.   This happened to get 'tested'
while our own zuul merges were not functioning, so it ended up getting
merged undetected and broke our gate.

This fixes the test case and moves the review states
we care about to constants, for easier updating across code and tests later.
It also tightens up the review handling to only process reviews with states
we care about, as the API is in dev-preview and may add new states in the
future.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>