Ensure PRs arent rejected for stale negative reviews
This updates the github source to only use the most recent review
from each user reviewing a PR.  This avoids having obsolete negative
reviews trump a newer positive review.

Since the timestamp in an approval's 'grantedOn' is used elsewhere,
I've added a new test for good measure to ensure it works for the
github case, where we are doing some conversion of github formatted
timestamps to internal timestamps.

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>