Catch creator is None.
A weird scenario, if a status exists for a user that has been deleted
this will show as "ghost" on the github web ui and put null into the
json of delivered status events.

This happened when I was testing integrations and had recreated the
integration and bot user whilst the old status was still visible.

Change-Id: If5088ca6a4aa13f4946a0718576ce88f115b18e4
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>