Install github3 from commit hash only in testing
pip install -U . doesn't respect installing from git, so in production
we are ending up with 0.9.6 being installed. This has worked in testing
because pip install -r does work.

We only install this in testing now so that we don't have versions
churning between the pip install zuul and install github3 points.

Change-Id: I5bf6ae532210aa37c2e89e4cb7dfcb8e4ec76e69
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>